reorg and consolidate rpc api modules
