+++ crates/sui-json-rpc/src/api/read.rs
@@ -1,0 +1,233 @@
+// Copyright (c) Mysten Labs, Inc.
+// SPDX-License-Identifier: Apache-2.0
+
+use fastcrypto::encoding::Base64;
+use jsonrpsee::core::RpcResult;
+use jsonrpsee_proc_macros::rpc;
+use std::collections::BTreeMap;
+use sui_json_rpc_types::{
+    Checkpoint, CheckpointId, DevInspectResults, DynamicFieldPage, GetObjectDataResponse,
+    GetPastObjectDataResponse, GetRawObjectDataResponse, MoveFunctionArgType,
+    SuiMoveNormalizedFunction, SuiMoveNormalizedModule, SuiMoveNormalizedStruct, SuiObjectInfo,
+    SuiTransactionAuthSignersResponse, SuiTransactionEffects, SuiTransactionResponse,
+    TransactionsPage,
+};
+use sui_open_rpc_macros::open_rpc;
+use sui_types::base_types::{
+    EpochId, ObjectID, SequenceNumber, SuiAddress, TransactionDigest, TxSequenceNumber,
+};
+use sui_types::digests::{CheckpointContentsDigest, CheckpointDigest};
+use sui_types::messages_checkpoint::{
+    CheckpointContents, CheckpointSequenceNumber, CheckpointSummary,
+};
+use sui_types::query::TransactionQuery;
+
+#[open_rpc(namespace = "sui", tag = "Read API")]
+#[rpc(server, client, namespace = "sui")]
+pub trait ReadApi {
+    /// Return the list of objects owned by an address.
+    #[method(name = "getObjectsOwnedByAddress")]
+    async fn get_objects_owned_by_address(
+        &self,
+        /// the owner's Sui address
+        address: SuiAddress,
+    ) -> RpcResult<Vec<SuiObjectInfo>>;
+
+    /// Return the list of dynamic field objects owned by an object.
+    #[method(name = "getDynamicFields")]
+    async fn get_dynamic_fields(
+        &self,
+        /// The ID of the parent object
+        parent_object_id: ObjectID,
+        /// Optional paging cursor
+        cursor: Option<ObjectID>,
+        /// Maximum item returned per page, default to [QUERY_MAX_RESULT_LIMIT] if not specified.
+        limit: Option<usize>,
+    ) -> RpcResult<DynamicFieldPage>;
+
+    /// Return the total number of transactions known to the server.
+    #[method(name = "getTotalTransactionNumber")]
+    async fn get_total_transaction_number(&self) -> RpcResult<u64>;
+
+    /// Return list of transaction digests within the queried range.
+    #[method(name = "getTransactionsInRange")]
+    async fn get_transactions_in_range(
+        &self,
+        /// the matching transactions' sequence number will be greater than or equals to the starting sequence number
+        start: TxSequenceNumber,
+        /// the matching transactions' sequence number will be less than the ending sequence number
+        end: TxSequenceNumber,
+    ) -> RpcResult<Vec<TransactionDigest>>;
+
+    /// Return the transaction response object.
+    #[method(name = "getTransaction")]
+    async fn get_transaction(
+        &self,
+        /// the digest of the queried transaction
+        digest: TransactionDigest,
+    ) -> RpcResult<SuiTransactionResponse>;
+
+    /// Return the authority public keys that commits to the authority signature of the transaction.
+    #[method(name = "getTransactionAuthSigners")]
+    async fn get_transaction_auth_signers(
+        &self,
+        /// the digest of the queried transaction
+        digest: TransactionDigest,
+    ) -> RpcResult<SuiTransactionAuthSignersResponse>;
+
+    /// Return the object information for a specified object
+    #[method(name = "getObject")]
+    async fn get_object(
+        &self,
+        /// the ID of the queried object
+        object_id: ObjectID,
+    ) -> RpcResult<GetObjectDataResponse>;
+
+    /// Return the dynamic field object information for a specified object
+    #[method(name = "getDynamicFieldObject")]
+    async fn get_dynamic_field_object(
+        &self,
+        /// The ID of the queried parent object
+        parent_object_id: ObjectID,
+        /// The Name of the dynamic field
+        name: String,
+    ) -> RpcResult<GetObjectDataResponse>;
+
+    /// Runs the transaction in dev-inspect mode. Which allows for nearly any
+    /// transaction (or Move call) with any arguments. Detailed results are
+    /// provided, including both the transaction effects and any return values.
+    #[method(name = "devInspectTransaction")]
+    async fn dev_inspect_transaction(
+        &self,
+        sender_address: SuiAddress,
+        /// BCS encoded TransactionKind(as opposed to TransactionData, which include gasBudget and gasPrice)
+        tx_bytes: Base64,
+        /// Gas is not charged, but gas usage is still calculated. Default to use reference gas price
+        gas_price: Option<u64>,
+        /// The epoch to perform the call. Will be set from the system state object if not provided
+        epoch: Option<EpochId>,
+    ) -> RpcResult<DevInspectResults>;
+
+    /// Return transaction execution effects including the gas cost summary,
+    /// while the effects are not committed to the chain.
+    #[method(name = "dryRunTransaction")]
+    async fn dry_run_transaction(&self, tx_bytes: Base64) -> RpcResult<SuiTransactionEffects>;
+
+    /// Return the argument types of a Move function,
+    /// based on normalized Type.
+    #[method(name = "getMoveFunctionArgTypes")]
+    async fn get_move_function_arg_types(
+        &self,
+        package: ObjectID,
+        module: String,
+        function: String,
+    ) -> RpcResult<Vec<MoveFunctionArgType>>;
+
+    /// Return structured representations of all modules in the given package
+    #[method(name = "getNormalizedMoveModulesByPackage")]
+    async fn get_normalized_move_modules_by_package(
+        &self,
+        package: ObjectID,
+    ) -> RpcResult<BTreeMap<String, SuiMoveNormalizedModule>>;
+
+    /// Return a structured representation of Move module
+    #[method(name = "getNormalizedMoveModule")]
+    async fn get_normalized_move_module(
+        &self,
+        package: ObjectID,
+        module_name: String,
+    ) -> RpcResult<SuiMoveNormalizedModule>;
+
+    /// Return a structured representation of Move struct
+    #[method(name = "getNormalizedMoveStruct")]
+    async fn get_normalized_move_struct(
+        &self,
+        package: ObjectID,
+        module_name: String,
+        struct_name: String,
+    ) -> RpcResult<SuiMoveNormalizedStruct>;
+
+    /// Return a structured representation of Move function
+    #[method(name = "getNormalizedMoveFunction")]
+    async fn get_normalized_move_function(
+        &self,
+        package: ObjectID,
+        module_name: String,
+        function_name: String,
+    ) -> RpcResult<SuiMoveNormalizedFunction>;
+
+    /// Return list of transactions for a specified query criteria.
+    #[method(name = "getTransactions")]
+    async fn get_transactions(
+        &self,
+        /// the transaction query criteria.
+        query: TransactionQuery,
+        /// Optional paging cursor
+        cursor: Option<TransactionDigest>,
+        /// Maximum item returned per page, default to [QUERY_MAX_RESULT_LIMIT] if not specified.
+        limit: Option<usize>,
+        /// query result ordering, default to false (ascending order), oldest record first.
+        descending_order: Option<bool>,
+    ) -> RpcResult<TransactionsPage>;
+
+    /// Note there is no software-level guarantee/SLA that objects with past versions
+    /// can be retrieved by this API, even if the object and version exists/existed.
+    /// The result may vary across nodes depending on their pruning policies.
+    /// Return the object information for a specified version
+    #[method(name = "tryGetPastObject")]
+    async fn try_get_past_object(
+        &self,
+        /// the ID of the queried object
+        object_id: ObjectID,
+        /// the version of the queried object. If None, default to the latest known version
+        version: SequenceNumber,
+    ) -> RpcResult<GetPastObjectDataResponse>;
+
+    /// Return the sequence number of the latest checkpoint that has been executed
+    #[method(name = "getLatestCheckpointSequenceNumber")]
+    fn get_latest_checkpoint_sequence_number(&self) -> RpcResult<CheckpointSequenceNumber>;
+
+    /// Return a checkpoint
+    #[method(name = "getCheckpoint")]
+    fn get_checkpoint(
+        &self,
+        /// Checkpoint identifier, can use either checkpoint digest, or checkpoint sequence number as input.
+        id: CheckpointId,
+    ) -> RpcResult<Checkpoint>;
+
+    /// Return a checkpoint summary based on a checkpoint sequence number
+    #[method(name = "getCheckpointSummary")]
+    fn get_checkpoint_summary(
+        &self,
+        sequence_number: CheckpointSequenceNumber,
+    ) -> RpcResult<CheckpointSummary>;
+
+    /// Return a checkpoint summary based on checkpoint digest
+    #[method(name = "getCheckpointSummaryByDigest")]
+    fn get_checkpoint_summary_by_digest(
+        &self,
+        digest: CheckpointDigest,
+    ) -> RpcResult<CheckpointSummary>;
+
+    /// Return contents of a checkpoint, namely a list of execution digests
+    #[method(name = "getCheckpointContents")]
+    fn get_checkpoint_contents(
+        &self,
+        sequence_number: CheckpointSequenceNumber,
+    ) -> RpcResult<CheckpointContents>;
+
+    /// Return contents of a checkpoint based on checkpoint content digest
+    #[method(name = "getCheckpointContentsByDigest")]
+    fn get_checkpoint_contents_by_digest(
+        &self,
+        digest: CheckpointContentsDigest,
+    ) -> RpcResult<CheckpointContents>;
+
+    /// Return the raw BCS serialized move object bytes for a specified object.
+    #[method(name = "getRawObject")]
+    async fn get_raw_object(
+        &self,
+        /// the id of the object
+        object_id: ObjectID,
+    ) -> RpcResult<GetRawObjectDataResponse>;
+}
