+++ crates/sui-json-rpc/src/error.rs
@@ -1,6 +1,7 @@
 // Copyright (c) Mysten Labs, Inc.
 // SPDX-License-Identifier: Apache-2.0
 
+use hyper::header::InvalidHeaderValue;
 use jsonrpsee::core::Error as RpcError;
 use jsonrpsee::types::error::CallError;
 use sui_types::error::SuiError;
@@ -8,7 +9,7 @@ use thiserror::Error;
 
 #[derive(Debug, Error)]
 pub enum Error {
-    #[error("Sui error: {0}")]
+    #[error(transparent)]
     SuiError(#[from] SuiError),
 
     #[error("{0}")]
@@ -19,6 +20,12 @@ pub enum Error {
 
     #[error("Unexpected error: {0}")]
     UnexpectedError(String),
+
+    #[error(transparent)]
+    RPCServerError(#[from] jsonrpsee::core::Error),
+
+    #[error(transparent)]
+    InvalidHeaderValue(#[from] InvalidHeaderValue),
 }
 
 impl From<Error> for RpcError {
