+++ crates/sui-json-rpc/tests/routing_tests.rs
@@ -25,7 +25,7 @@ use sui_open_rpc_macros::open_rpc;
 
 #[tokio::test]
 async fn test_rpc_backward_compatibility() {
-    let mut builder = JsonRpcServerBuilder::new("1.5", &Registry::new()).unwrap();
+    let mut builder = JsonRpcServerBuilder::new("1.5", &Registry::new());
     builder.register_module(TestApiModule).unwrap();
 
     let port = get_available_port("0.0.0.0");
@@ -109,7 +109,7 @@ async fn test_rpc_backward_compatibility() {
 async fn test_disable_routing() {
     env::set_var("DISABLE_BACKWARD_COMPATIBILITY", "true");
 
-    let mut builder = JsonRpcServerBuilder::new("1.5", &Registry::new()).unwrap();
+    let mut builder = JsonRpcServerBuilder::new("1.5", &Registry::new());
     builder.register_module(TestApiModule).unwrap();
 
     let port = get_available_port("0.0.0.0");
@@ -148,7 +148,7 @@ async fn test_disable_routing() {
 
 #[tokio::test]
 async fn test_rpc_backward_compatibility_batched_request() {
-    let mut builder = JsonRpcServerBuilder::new("1.5", &Registry::new()).unwrap();
+    let mut builder = JsonRpcServerBuilder::new("1.5", &Registry::new());
     builder.register_module(TestApiModule).unwrap();
 
     let port = get_available_port("0.0.0.0");
